Our first functioning auto
Modifying the gobilda starter code to incorporate mecanum wheels and remove encoders. (Our robot won’t have encoder cables installed in time for the first competition.)

The code is expecting the robot to be up at the front of the goal. It launches three artifacts, backs up, rotates, strafes and backs up again to be out of the launch zone.

It’s only been tested on the red goal and only on a non-official field setup, so we will need to fine tune at the competition.