Added indicators to ensure proper auto init
I added a few more LEDs to the robot.
During the auto init, the LEDs show the operators whether the Yaw is near zero and which alliance they have chosen.
During the primary loop, I want the second LEDs to specify whether the robot is both in the correct position and the flywheel is rotating at the correct velocity to reliably shoot a basket.