Tuned PIDF and launcher velocities for shooting from back zone
We changed the motor from a ~300RPM motor to a 6000 RPM motor and added a hood. The values adjusted here are set after tuning to reliable shoot from the back launch zone.

Also - we added another set of LEDs which we will use to indicate to the driver when the launch more artifacts.