Adjust auto mode for shooting from back launch zone
Adjust the auto program from shooting right under the basket to shooting from the back launch zone. The robot needs to be placed in the launch tile either on the left or the right side depending on whether the alliance is red or blue. Then, it will move, shoot and end up the loading zone ready for teleop.