Removed sensors from opmode for competition
Before the last competition, we were hoping that sensors should help us align the robot and make shooting more accurate. We couldn’t get any of that to work, so we removed all that code, mostly by commenting it out.

As I’m about to embark on a mission to refactor this, I wanted to check it in first.